// src/modules/storefront/interfaces/Theme.ts
export interface DynamicApiModuleInterface {
  regexp: any;
  matches: any;
  groups?: any;
  apis: {
    apiName: string;
    url: string;
    strapiLocalization?: boolean;
  }[];
}